requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
gunicorn==21.2.0